Allow provisioning of "default" github/transifex credentials via environment variables
Having the credentials in the resource files makes it very annoying and dangerous
to keep this project in a public repository such as GitHub, so with the environment
variables it should be possible to remove the credentials from the configuration.

Note that this also means one cannot have different users for each project, but this
is something we could potentially fix later on (by actually killing the configuration,
and going for a system similar to travis, which keeps a database of the authorized projects).