"Fork" the project properly
We will do quite a few changes to this project, and the initial project
seems to be more or less unmaintained. The plan is still to feed back as
much as possible back though, but it seems unrealistic at this point, so
we will anyways have to maintain our own version.